feat: update front with all energies models
